Corrette funzionalita' promemoria

+++ resources/views/promemoria/_elenco.blade.php
@@ -12,7 +12,9 @@
                 <th>Chi</th>
                 <th>Quando</th>
                 <th>Cosa</th>
-                <th>&nbsp;</th>
+                @can('modificare-agenda', $pratica)
+                    <th>&nbsp;</th>
+                @endcan
             </thead>
             <tbody>
                 @foreach ($promemoria as $p)
@@ -20,14 +22,17 @@
                         <td class="table-text col-md-2"><div>{{ $p->chi }}</div></td>
                         <td class="table-text col-md-2"><div>{{ date_diff_days($p->quando) }}</div></td>
                         <td class="table-text col-md-6"><div>{{ $p->cosa }}</div></td>
-                        <td class="col-md-2">
-                            {!! Form::open(['action' => ['PromemoriaController@destroy', 'cliente' => $p->pratica->cliente,
-                                'pratica' => $p->pratica, 'promemoria' => $p], 'method' => 'delete']) !!}
-                                <button type="submit" class="btn btn-success">
-                                    <i class="fa fa-fw fa-check"></i>
-                                </button>
-                            {!! Form::close() !!}
-                        </td>
+                        
+                        @can('modificare-agenda', $pratica)
+                            <td class="col-md-2">
+                                {!! Form::open(['action' => ['PromemoriaController@destroy', 'cliente' => $p->pratica->cliente,
+                                    'pratica' => $p->pratica, 'promemoria' => $p], 'method' => 'delete']) !!}
+                                    <button type="submit" class="btn btn-success">
+                                        <i class="fa fa-fw fa-check"></i>
+                                    </button>
+                                {!! Form::close() !!}
+                            </td>
+                        @endcan
                     </tr>
                 @endforeach
             </tbody>
@@ -36,10 +41,12 @@
             <p class="text-center">Non sono presenti promemoria da completare per questa pratica</p>
         @endif
         <div class="text-center">
-            <button type="button" class="btn btn-success" data-toggle="modal" data-target="#createPromemoriaModal">
-                <i class="fa fa-fw fa-plus"></i>
-                Aggiungi promemoria
-            </button>
+            @can('modificare-agenda', $pratica)
+                <button type="button" class="btn btn-success" data-toggle="modal" data-target="#createPromemoriaModal">
+                    <i class="fa fa-fw fa-plus"></i>
+                    Aggiungi promemoria
+                </button>
+            @endcan
         </div>
     </div>
 </div>
